+++ Telescope/DQM/headers/Ccorrel_line_finder.h
@@ -22,27 +22,27 @@ class Ccorrel_line_finder{
 private:
 	TH2F*					_plot;
 	Ccorrel_line*			_line;
-	float					_theta_lower;
-	float					_theta_upper;
+	double					_theta_lower;
+	double					_theta_upper;
 	std::vector<TH1F*>		_projections;
 	std::vector<TH1F*>		_capped_projections;
 	std::vector<TH1F*>		_signal_projections;
 	int 					_n_projections;
 	TGraph*					_stationary_graph;
 	TCanvas*				_c1;
-	float					_temp_theta;
+	double					_temp_theta;
 	int						_nproj_bins;
 	bool					_save_switch;
 	TDirectory*				_save_dir;
 
 
 public:
-	Ccorrel_line_finder(TH2F*, float, float, int, bool, TDirectory*);
+	Ccorrel_line_finder(TH2F*, double, double, int, bool, TDirectory*);
 	~Ccorrel_line_finder();
 	void					fill_projections();
 	void 					varying_projections();
 	void					fill_stationary_plot();
-	float					find_height_stationary_parameter(int, int&);
+	double					find_height_stationary_parameter(int, int&);
 	void					fill_signal_projection(int);
 	void					save_figs();
 	void					set_line();
@@ -53,11 +53,11 @@ public:
 	TH2F*					get_plot(){return _plot;}
 
 
-	void					set_theta_lower(float t){_theta_lower = t;}
-	float					get_theta_lower(){return _theta_lower;}
+	void					set_theta_lower(double t){_theta_lower = t;}
+	double					get_theta_lower(){return _theta_lower;}
 
-	void					set_theta_upper(float t){_theta_upper = t;}
-	float					get_theta_upper(){return _theta_upper;}
+	void					set_theta_upper(double t){_theta_upper = t;}
+	double					get_theta_upper(){return _theta_upper;}
 
 	Ccorrel_line*			get_line(){return _line;};
